Update serverToClientConfig and clientToServerConfig

+++ TicTacToeGame.Client/Game/CellFactory.cs
@@ -5,7 +5,9 @@ namespace TicTacToeGame.Client.Game
 {
     public static class CellFactory
     {
-        public static List<BoardCell> Build(ushort count) => 
-            Enumerable.Range(0, count).Select(index => new BoardCell(index)).ToList();
+        public static List<BoardCell> Build(ushort count) =>
+            Enumerable.Range(0, count)
+                      .Select(index => new BoardCell((ushort)index))
+                      .ToList();
     }
 }
